test: optimize tests by adding shared utility functions
