+++ tests/patterns.spec.ts
@@ -1,32 +1,10 @@
-import { test, expect } from '@playwright/test'
-import moment from 'moment'
+import { test } from '@playwright/test'
+import { verifyApiCallToGtfsAgenciesList, verifyDateFromParameter } from '../tests/utils.js'
 
-test('verify API call to gtfs_agencies/list - "Patterns"', async ({ page }) => {
-  let apiCallMade = false
-  page.on('request', (request) => {
-    if (request.url().includes('gtfs_agencies/list')) {
-      apiCallMade = true
-    }
-  })
-
-  await page.goto('/')
-  await page.getByRole('link', { name: 'דפוסי נסיעות שלא בוצעו' }).click()
-  await page.getByLabel('חברה מפעילה').click()
-  expect(apiCallMade).toBeTruthy()
+test('verify API call to gtfs_agencies/list - "patterns"', async ({ page }) => {
+  await verifyApiCallToGtfsAgenciesList(page, 'דפוסי נסיעות שלא בוצעו')
 })
 
-test('Verify date_from parameter from "Patterns"', async ({ page }) => {
-  const apiRequest = page.waitForRequest((request) => request.url().includes('gtfs_agencies/list'))
-
-  await page.goto('/')
-  await page.getByRole('link', { name: 'דפוסי נסיעות שלא בוצעו' }).click()
-
-  const request = await apiRequest
-  const url = new URL(request.url())
-  const dateFromParam = url.searchParams.get('date_from')
-  const dateFrom = moment(dateFromParam)
-  const daysAgo = moment().diff(dateFrom, 'days')
-
-  expect(daysAgo).toBeGreaterThanOrEqual(0)
-  expect(daysAgo).toBeLessThanOrEqual(3)
+test('Verify date_from parameter from "patterns"', async ({ page }) => {
+  await verifyDateFromParameter(page, 'דפוסי נסיעות שלא בוצעו')
 })
