+++ tests/singlelineTest.spec.ts
@@ -1,5 +1,5 @@
-import moment from 'moment'
 import SinglelinePage from '../src/test_pages/SinglelinePage'
+import { verifyApiCallToGtfsAgenciesList, verifyDateFromParameter } from '../tests/utils.js'
 import { getPastDate, test, expect, urlMatcher } from './utils'
 
 test.describe('Single line page tests', () => {
@@ -100,31 +100,9 @@ test.describe('Single line page tests', () => {
 })
 
 test('verify API call to gtfs_agencies/list - "Map by line"', async ({ page }) => {
-  let apiCallMade = false
-  page.on('request', (request) => {
-    if (request.url().includes('gtfs_agencies/list')) {
-      apiCallMade = true
-    }
-  })
-
-  await page.goto('/')
-  await page.getByRole('link', { name: 'מפה לפי קו' }).click()
-  await page.getByLabel('חברה מפעילה').click()
-  expect(apiCallMade).toBeTruthy()
+  await verifyApiCallToGtfsAgenciesList(page, 'מפה לפי קו')
 })
 
 test('Verify date_from parameter from "Map by line"', async ({ page }) => {
-  const apiRequest = page.waitForRequest((request) => request.url().includes('gtfs_agencies/list'))
-
-  await page.goto('/')
-  await page.getByRole('link', { name: 'מפה לפי קו' }).click()
-
-  const request = await apiRequest
-  const url = new URL(request.url())
-  const dateFromParam = url.searchParams.get('date_from')
-  const dateFrom = moment(dateFromParam)
-  const daysAgo = moment().diff(dateFrom, 'days')
-
-  expect(daysAgo).toBeGreaterThanOrEqual(0)
-  expect(daysAgo).toBeLessThanOrEqual(3)
+  await verifyDateFromParameter(page, 'מפה לפי קו')
 })
