+++ tests/tripExistence.spec.ts
@@ -1,4 +1,4 @@
-import moment from 'moment'
+import { verifyApiCallToGtfsAgenciesList, verifyDateFromParameter } from '../tests/utils.js'
 import { test, expect, urlMatcher } from './utils'
 
 test.describe('Trip Existence Page Tests', () => {
@@ -34,33 +34,9 @@ test.describe('Trip Existence Page Tests', () => {
 })
 
 test('verify API call to gtfs_agencies/list - "Planned trips"', async ({ page }) => {
-  let apiCallMade = false
-  page.on('request', (request) => {
-    if (request.url().includes('gtfs_agencies/list')) {
-      apiCallMade = true
-    }
-  })
-
-  await page.goto('/')
-  await page.getByRole('link', { name: 'קיום נסיעות' }).click()
-  await page.getByLabel('חברה מפעילה').click()
-  expect(apiCallMade).toBeTruthy()
+  await verifyApiCallToGtfsAgenciesList(page, 'קיום נסיעות')
 })
 
-test('the dateFrom parameter should be recent when visiting the "Planned trips" tabs', async ({
-  page,
-}) => {
-  const apiRequest = page.waitForRequest((request) => request.url().includes('gtfs_agencies/list'))
-
-  await page.goto('/')
-  await page.getByRole('link', { name: 'קיום נסיעות' }).click()
-
-  const request = await apiRequest
-  const url = new URL(request.url())
-  const dateFromParam = url.searchParams.get('date_from')
-  const dateFrom = moment(dateFromParam)
-  const daysAgo = moment().diff(dateFrom, 'days')
-
-  expect(daysAgo).toBeGreaterThanOrEqual(0)
-  expect(daysAgo).toBeLessThanOrEqual(3)
+test('Verify date_from parameter from "planned trips"', async ({ page }) => {
+  await verifyDateFromParameter(page, 'קיום נסיעות')
 })
